requirements.txt
Flask
jinja2
flask_pymongo

